Fix active arg parsing
arg is the string with the value passed for the active effect. Calling bool() on a non-empty string always returns True. As a result calling `razer-cli -e active,0` sets the device as active instead of inactive.
This commit fixes that by converting the arg to an integer before calling bool() on it.